Refactor typingsInstaller helpers
